updated - scheduled backup and task fix

+++ functions/command_autorestart.ps1
@@ -11,27 +11,32 @@ Function New-RestartJobBG {
     If ($env:UserName -and $env:COMPUTERNAME) {
         $UserName = "$env:COMPUTERNAME\$env:UserName"
         Write-Host "Run Task Whether user is logged on or not" -F Cyan -NoNewline
-        Write-Host "Input AutoRestart Time. ie 3am: " -F Cyan -NoNewline
+        Write-Host "Input AutoRestart Time. i.e. 3am: " -F Cyan -NoNewline
         $restartTime = Read-Host
         If ($restartTime) {
             Write-Host "Username: $env:COMPUTERNAME\$env:UserName"
-            $SecurePassword = $password = Read-Host "Password:" -AsSecureString
-            If ($UserName -and $SecurePassword) {
-                $Credentials = New-Object System.Management.Automation.PSCredential -ArgumentList $UserName, $SecurePassword
-                $Password = $Credentials.GetNetworkCredential().Password  
-                $Action = New-ScheduledTaskAction -Execute 'powershell.exe' -Argument "$currentdir\ssm.ps1 restart $serverfiles" -WorkingDirectory "$currentdir"
-                #$Trigger = New-ScheduledTaskTrigger -Once -At (Get-Date).Date -RepetitionInterval (New-TimeSpan -Minutes $restartTime)
-                $Trigger = New-ScheduledTaskTrigger -Daily -At $restartTime
-                $Settings = New-ScheduledTaskSettingsSet -ExecutionTimeLimit '00:00:00'
-                $Task = New-ScheduledTask -Action $Action -Trigger $Trigger -Settings $Settings
-                Get-Infomessage "Creating Task" 'start'
-                Register-ScheduledTask -TaskName "$serverfiles $command" -InputObject $Task -User "$UserName" -Password "$Password" -ea SilentlyContinue | Out-File -Append -Encoding Default  $ssmlog
-                If ($?){
-                    clear-hostline 1
-                    Get-Infomessage "Creating Task" 'done'
-                } Else{
-                    clear-hostline 1
-                    Get-Infomessage "Creating Task" $false
+            $SecurePassword = Read-Host "Password:" -AsSecureString
+            # C:\Windows\System32\WindowsPowerShell\v1.0\powershell.exe
+            If ((Test-Path "C:\Windows\System32\WindowsPowerShell\v1.0\powershell.exe")) {
+                $posh = "C:\Windows\System32\WindowsPowerShell\v1.0\powershell.exe"
+                If ($UserName -and $SecurePassword) {
+                    $Credentials = New-Object System.Management.Automation.PSCredential -ArgumentList $UserName, $SecurePassword
+                    $Password = $Credentials.GetNetworkCredential().Password  
+                    $Action = New-ScheduledTaskAction -Execute "$posh" -Argument "$currentdir\ssm.ps1 restart $serverfiles" -WorkingDirectory "$currentdir"
+                    #$Trigger = New-ScheduledTaskTrigger -Once -At (Get-Date).Date -RepetitionInterval (New-TimeSpan -Minutes $restartTime)
+                    $Trigger = New-ScheduledTaskTrigger -Daily -At $restartTime
+                    $Settings = New-ScheduledTaskSettingsSet -ExecutionTimeLimit '00:00:00'
+                    $Task = New-ScheduledTask -Action $Action -Trigger $Trigger -Settings $Settings
+                    Get-Infomessage "Creating Task" 'start'
+                    Register-ScheduledTask -TaskName "$serverfiles $command" -InputObject $Task -User "$UserName" -Password "$Password" -ea SilentlyContinue | Out-File -Append -Encoding Default  $ssmlog
+                    If ($?) {
+                        clear-hostline 1
+                        Get-Infomessage "Creating Task" 'done'
+                    }
+                    Else {
+                        clear-hostline 1
+                        Get-Infomessage "Creating Task" $false
+                    }
                 }
             }
         }
@@ -43,19 +48,23 @@ Function New-RestartJob {
     Write-Host "Input AutoRestart Time. ie 3am: " -F Cyan -NoNewline
     $restartTime = Read-Host
     If ($restartTime) {
-        $Action = New-ScheduledTaskAction -Execute 'powershell.exe' -Argument "$currentdir\ssm.ps1 restart $serverfiles" -WorkingDirectory "$currentdir"
-        #$Trigger = New-ScheduledTaskTrigger -Once -At (Get-Date).Date -RepetitionInterval (New-TimeSpan -Hours $restartTime)
-        $Trigger = New-ScheduledTaskTrigger -Daily -At $restartTime
-        $Settings = New-ScheduledTaskSettingsSet -ExecutionTimeLimit '00:00:00'
-        $Task = New-ScheduledTask -Action $Action -Trigger $Trigger -Settings $Settings
-        Get-Infomessage "Creating Task" 'start'
-        Register-ScheduledTask -TaskName "$serverfiles $command" -InputObject $Task | Out-File -Append -Encoding Default  $ssmlog
-        If ($?){
-            clear-hostline 1
-            Get-Infomessage "Creating Task" 'done'
-        } Else{
-            clear-hostline 1
-            Get-Infomessage "Creating Task" $false
+        If ((Test-Path "C:\Windows\System32\WindowsPowerShell\v1.0\powershell.exe")) {
+            $posh = "C:\Windows\System32\WindowsPowerShell\v1.0\powershell.exe"
+            $Action = New-ScheduledTaskAction -Execute "$posh" -Argument "$currentdir\ssm.ps1 restart $serverfiles" -WorkingDirectory "$currentdir"
+            #$Trigger = New-ScheduledTaskTrigger -Once -At (Get-Date).Date -RepetitionInterval (New-TimeSpan -Hours $restartTime)
+            $Trigger = New-ScheduledTaskTrigger -Daily -At $restartTime
+            $Settings = New-ScheduledTaskSettingsSet -ExecutionTimeLimit '00:00:00'
+            $Task = New-ScheduledTask -Action $Action -Trigger $Trigger -Settings $Settings
+            Get-Infomessage "Creating Task" 'start'
+            Register-ScheduledTask -TaskName "$serverfiles $command" -InputObject $Task | Out-File -Append -Encoding Default  $ssmlog
+            If ($?) {
+                clear-hostline 1
+                Get-Infomessage "Creating Task" 'done'
+            }
+            Else {
+                clear-hostline 1
+                Get-Infomessage "Creating Task" $false
+            }
         }
     }
 }
